requirements.txt
streamlit==0.80.0
pandas==1.2.4
Pillow==8.2.0